Changes WeeklyGoal to Goal  and adds a strengthValue attribute
