+++ HealthChallenge/Model/Goal.swift
@@ -1,0 +1,50 @@
+//
+//  WeekGoal.swift
+//  HealthChallenge
+//
+//  Created by Dominic Lanzillotta on 3/12/19.
+//  Copyright © 2019 Dominic Lanzillotta. All rights reserved.
+//
+
+import Foundation
+import CloudKit
+
+class Goal {
+    var name: String
+    var recordID: CKRecord.ID
+    var isPublic: Bool
+    var reviewForPublic: Bool
+    var strengthValue: Int
+    
+    init(name: String, isPublic: Bool = false, reviewForPublic: Bool = false, strengthValue: Int = 1,recordID: CKRecord.ID = CKRecord.ID(recordName: UUID().uuidString)) {
+        self.name = name
+        self.isPublic = isPublic
+        self.reviewForPublic = reviewForPublic
+        self.recordID = recordID
+        self.strengthValue = strengthValue
+    }
+    
+    init?(record: CKRecord) {
+        guard let name = record[Goal.nameKey] as? String,
+            let isPublic = record[Goal.isPublicKey] as? Bool,
+            let strengthValue = record[Goal.strengthValueKey] as? Int,
+            let reviewForPublic = record[Goal.reviewForPublicKey] as? Bool else {return nil}
+        
+        self.name = name
+        self.recordID = record.recordID
+        self.isPublic = isPublic
+        self.reviewForPublic = reviewForPublic
+        self.strengthValue = strengthValue
+
+    }
+}
+extension CKRecord {
+    
+    convenience init(goal: Goal) {
+        self.init(recordType: Goal.typeKey, recordID: goal.recordID)
+        
+        self.setValue(goal.name, forKey: Goal.nameKey)
+        self.setValue(goal.isPublic, forKey: Goal.isPublicKey)
+        self.setValue(goal.reviewForPublic, forKey: Goal.reviewForPublicKey)
+    }
+}
