--- HealthChallenge/Model/MagicStrings/MonthlyGoal+Extension.swift
@@ -1,13 +1,0 @@
-//
-//  MonthlyGoal+Extension.swift
-//  HealthChallenge
-//
-//  Created by Dominic Lanzillotta on 3/12/19.
-//  Copyright © 2019 Dominic Lanzillotta. All rights reserved.
-//
-
-import Foundation
-
-extension MonthlyGoal {
-
-}
